backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
pymongo==4.5.0
Werkzeug==2.3.7
reportlab==4.0.4
tldextract==3.6.0
python-socketio==5.9.0
eventlet==0.33.3
psutil==5.9.6
asyncio==3.4.3
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4